feat: update application to use 'locations' instead of 'stores'
- Refactored the proxy redirection logic to redirect users to '/admin/locations' and '/user/locations' instead of the previous '/admin/stores' and '/user/stores'.
- Updated various components, including AdminPage and UserPage, to reflect the new 'locations' terminology.
- Introduced new pages and components for managing locations, including content and detail views.
- Enhanced user interface elements to align with the updated navigation structure and terminology.
