+++ frontend/src/app/admin/locations/[id]/loading.tsx
