+++ frontend/src/app/admin/locations/_old/content.tsx
