+++ frontend/src/app/admin/locations/page.tsx
