+++ frontend/src/app/admin/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from 'next/navigation';
 
 export default async function AdminPage() {
-  redirect('/admin/stores');
+  redirect('/admin/locations');
 }
