+++ frontend/src/app/user/locations/[id]/loading.tsx
