+++ frontend/src/app/user/locations/page.tsx
